refactor: introduce layers in the architecture
