+++ Controllers/OrderController.cs
@@ -1,6 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
+using TixTrack.WebApiInterview.Dtos;
+using TixTrack.WebApiInterview.Entities;
+using TixTrack.WebApiInterview.Repositories;
 
-namespace Tixtrack.WebApiInterview;
+namespace TixTrack.WebApiInterview.Controllers;
 
 [ApiController]
 [Route("[controller]")]
@@ -14,18 +17,18 @@ public class OrderController : ControllerBase
     [Route("")]
     public IEnumerable<Order> GetAll()
     {
-        var orders = Database.GetAllOrders();
+        var orders = OrderRepository.GetAllOrders();
         foreach (var o in orders)
         {
             if (o.Product1Id != null)
             {
-                Product product = Database.GetProduct(o.Product1Id);
+                Product product = ProductRepository.GetProduct(o.Product1Id);
                 o.Product1Name = product.Name;
                 o.Product1Price = product.Price;
             }
             if (o.Product2Id != null)
             {
-                Product product = Database.GetProduct(o.Product2Id);
+                Product product = ProductRepository.GetProduct(o.Product2Id);
                 o.Product2Name = product.Name;
                 o.Product2Price = product.Price;
             }
@@ -37,19 +40,19 @@ public class OrderController : ControllerBase
     [Route("{orderId}")]
     public Order? GetById(int orderId)
     {
-        var orders = Database.GetAllOrders();
+        var orders = OrderRepository.GetAllOrders();
         var order = orders.SingleOrDefault(o => o.Id == orderId);
         if (order != null)
         {
             if (order.Product1Id != null)
             {
-                Product product = Database.GetProduct(order.Product1Id);
+                Product product = ProductRepository.GetProduct(order.Product1Id);
                 order.Product1Name = product.Name;
                 order.Product1Price = product.Price;
             }
             if (order.Product2Id != null)
             {
-                Product product = Database.GetProduct(order.Product2Id);
+                Product product = ProductRepository.GetProduct(order.Product2Id);
                 order.Product2Name = product.Name;
                 order.Product2Price = product.Price;
             }
@@ -61,21 +64,15 @@ public class OrderController : ControllerBase
     [Route("")]
     public int Create(Order order)
     {
-        Database.CreateOrder(order);
+        OrderRepository.CreateOrder(order);
         return order.Id;
-    }
-
-    public class SalesReport
-    {
-        public int OrderCount { get; set; }
-        public double TotalSales { get; set; }
     }
 
     [HttpGet]
     [Route("salesreport")]
     public SalesReport GetSalesReport()
     {
-        IList<Order> orders = Database.GetAllOrders();
+        IList<Order> orders = OrderRepository.GetAllOrders();
         int OrderCount = 0;
         double totalSales = 0;
         foreach (var o in orders)
@@ -83,12 +80,12 @@ public class OrderController : ControllerBase
             OrderCount++;
             if (o.Product1Id != null)
             {
-                Product product = Database.GetProduct(o.Product1Id);
+                Product product = ProductRepository.GetProduct(o.Product1Id);
                 totalSales += (product.Price * o.Product1Quantiity!.Value);
             }
             if (o.Product2Id != null)
             {
-                Product product = Database.GetProduct(o.Product2Id);
+                Product product = ProductRepository.GetProduct(o.Product2Id);
                 totalSales += (product.Price * o.Product2Quantiity!.Value);
             }
         }
@@ -100,4 +97,3 @@ public class OrderController : ControllerBase
         };
     }
 }
-
