+++ Dtos/SalesReport.cs
@@ -1,0 +1,7 @@
+﻿namespace TixTrack.WebApiInterview.Dtos;
+
+public class SalesReport
+{
+    public int OrderCount { get; set; }
+    public double TotalSales { get; set; }
+}
