+++ Entities/Product.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace TixTrack.WebApiInterview.Entities;
+
+public class Product
+{
+    [DatabaseGenerated(DatabaseGeneratedOption.None)]
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public int AvailabileQuantity { get; set; }
+    public double Price { get; set; }
+    public string Type { get; set; } = string.Empty;
+}
