+++ Program.cs
@@ -1,12 +1,14 @@
-using Tixtrack.WebApiInterview;
+using TixTrack.WebApiInterview.Repositories;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
-
 builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
+
+if (builder.Environment.IsDevelopment())
+{
+    builder.Services.AddEndpointsApiExplorer();
+    builder.Services.AddSwaggerGen();
+}
 
 var app = builder.Build();
 
@@ -16,7 +18,9 @@ if (app.Environment.IsDevelopment())
     app.UseSwaggerUI();
 }
 
-Database.Seed();
+var db = new ApplicationContext();
+ProductRepository.Seed(db);
+OrderRepository.Seed(db);
 
 app.MapControllers();
 app.Run();
